src/requirements.txt
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
colorama==0.4.4
cymem==2.0.5
google-api-core==1.31.0
google-auth==1.33.0
googleapis-common-protos==1.53.0
idna==3.2
Jinja2==3.0.1
markovify==0.9.3
MarkupSafe==2.0.1
murmurhash==1.0.5
numpy==1.21.1
oauthlib==3.1.1
opencensus==0.7.13
opencensus-context==0.1.2
opencensus-ext-azure==1.0.8
packaging==21.0
pathy==0.6.0
pl-core-news-md @ https://github.com/explosion/spacy-models/releases/download/pl_core_news_md-3.1.0/pl_core_news_md-3.1.0-py3-none-any.whl
preshed==3.0.5
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==2.4.7
PySocks==1.7.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
schedule==1.1.0
six==1.16.0
smart-open==5.1.0
spacy==3.1.0
spacy-legacy==3.0.8
srsly==2.4.1
thinc==8.0.8
tqdm==4.61.2
tweepy==3.10.0
typer==0.3.2
typing-extensions==3.10.0.0
Unidecode==1.2.0
urllib3==1.26.6
wasabi==0.8.2
